Issues starting up resolved. Added migrations

+++ MediApp/Domain/Entities/County.cs
@@ -15,7 +15,7 @@ namespace Domain.Entities
         }
         public int Id { get; set; }
         public string Name { get; set; }
-        public int CountryId { get; set; }
+        public short CountryId { get; set; }
 
         public Country Country { get; set; }
         public ICollection<City> Cities { get; private set; }
